Fixed PKGeocoder
It will now serve consecutive items that could be served from cache. Turns out setting a property in its observer (in this instance didSet) does not trigger its observer again. The inconsistency comes when we use a completionHandler in the other code path where the result is fetched. This makes the bug hard to debug.

The solution is to delegate the queue popping to a function, where setting the property there would trigger the observer.

It should be noted that this is not the first time I was blocked by this issue (or feature)